% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spanel.R
\name{hausman}
\alias{hausman}
\title{Hausman test}
\usage{
hausman(fixed, random)
}
\arguments{
\item{fixed}{is the fixed effect object function}

\item{random}{is the random effect object function}
}
\value{
Chisq the hausman statistic

P-value the probability value

df the degree of freedom
}
\description{
Hausman test
}
\examples{
data(Produc)
data("usaww")
#fit the fixed function
fx<-span(log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp,Produc,usaww,n=48,t=17,model="fe")
# fit the random function
 ran<-span(log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp,Produc,usaww,n=48,t=17,model="re")
# the Hausman test
hausman(fx,ran)
}

