% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spanel.R
\name{span.formula}
\alias{span.formula}
\title{formula}
\usage{
\method{span}{formula}(formula, data = list(), w, n, t, model = c("fe",
  "be", "re"), ...)
}
\arguments{
\item{formula}{log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp}

\item{data}{the dataframe}

\item{w}{is the contiguity matrix}

\item{n}{the number of section}

\item{t}{the time per section}

\item{model}{"fe" for fixed effect "be" for between and "re" for random effect}

\item{...}{not used}
}
\description{
formula
}

