% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spanel.R
\name{summary.span}
\alias{summary.span}
\title{Summary}
\usage{
\method{summary}{span}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\description{
Summary
}

